+++ package.json
@@ -2,7 +2,10 @@
   "name": "vuepress-jsdoc",
   "version": "0.1.2",
   "description": "Generate jsdoc markdown files for vuepress",
-  "main": "index.js",
+  "entry": "index.js",
+  "bin": {
+    "vuepress-jsdoc": "bin/vuepress-jsdoc"
+  },
   "scripts": {
     "test": "jest",
     "test-coverage": "cross-env NODE_ENV=test jest --coverage",
@@ -47,9 +50,6 @@
     ],
     "coverageDirectory": "./coverage"
   },
-  "bin": {
-    "vuepress-jsdoc": "./index.js"
-  },
   "dependencies": {
     "@vuedoc/md": "^1.4.1",
     "bluebird": "^3.5.1",
@@ -58,7 +58,8 @@
     "husky": "^0.14.3",
     "jsdoc-to-markdown": "^4.0.1",
     "mkdirp": "^0.5.1",
-    "rimraf": "^2.6.2"
+    "rimraf": "^2.6.2",
+    "yargs": "^12.0.5"
   },
   "devDependencies": {
     "jest": "^22.4.4",
